+++ Lib/fontbakery/profiles/universal.py
@@ -7,7 +7,8 @@ from fontbakery.message import Message
 from fontbakery.fonts_profile import profile_factory
 from fontbakery.profiles.opentype import OPENTYPE_PROFILE_CHECKS
 
-profile_imports = ('fontbakery.profiles.opentype',)
+profile_imports = ('fontbakery.profiles.opentype',
+                   '.shared_conditions')
 profile = profile_factory(default_section=Section("Universal"))
 
 THIRDPARTY_CHECKS = [
@@ -34,6 +35,7 @@ UNIVERSAL_PROFILE_CHECKS = \
   'com.google.fonts/check/valid_glyphnames',
   'com.google.fonts/check/unique_glyphnames',
 #  'com.google.fonts/check/glyphnames_max_length',
+  'com.google.fonts/check/family/vertical_metrics',
   'com.google.fonts/check/superfamily/vertical_metrics',
 ]
 
@@ -723,9 +725,58 @@ def com_google_fonts_check_ttx_roundtrip(font):
 
 
 @check(
+  id = 'com.google.fonts/check/family/vertical_metrics',
+  rationale = """
+    We want all fonts within a family to have the same vertical metrics so their line spacing is consistent across the family.
+  """,
+  misc_metadata = {
+    'request': 'https://github.com/googlefonts/fontbakery/issues/1487'
+  }
+)
+def com_google_fonts_check_family_vertical_metrics(ttFonts):
+  """Each font in a family must have the same vertical metrics values."""
+  failed = []
+  vmetrics = {
+    "sTypoAscender": {},
+    "sTypoDescender": {},
+    "sTypoLineGap": {},
+    "usWinAscent": {},
+    "usWinDescent": {},
+    "ascent": {},
+    "descent": {}
+  }
+
+  for ttFont in ttFonts:
+    full_font_name = ttFont['name'].getName(4, 3, 1, 1033).toUnicode()
+    vmetrics['sTypoAscender'][full_font_name] = ttFont['OS/2'].sTypoAscender
+    vmetrics['sTypoDescender'][full_font_name] = ttFont['OS/2'].sTypoDescender
+    vmetrics['sTypoLineGap'][full_font_name] = ttFont['OS/2'].sTypoLineGap
+    vmetrics['usWinAscent'][full_font_name] = ttFont['OS/2'].usWinAscent
+    vmetrics['usWinDescent'][full_font_name] = ttFont['OS/2'].usWinDescent
+    vmetrics['ascent'][full_font_name] = ttFont['hhea'].ascent
+    vmetrics['descent'][full_font_name] = ttFont['hhea'].descent
+
+  for k, v in vmetrics.items():
+    metric_vals = set(vmetrics[k].values())
+    if len(metric_vals) != 1:
+      failed.append(k)
+
+  if failed:
+    for k in failed:
+      s = ["{}: {}".format(k, v) for k, v in vmetrics[k].items()]
+      s = "\n".join(s)
+      yield FAIL, (f"{k} is not the same across the family:\n"
+                   f"{s}")
+  else:
+    yield PASS, f"Vertical metrics are the same across the family."
+
+
+@check(
   id = 'com.google.fonts/check/superfamily/vertical_metrics',
   rationale = """
-    We want all fonts within a family (and eventual sibling families) to have the same vertical metrics so their line spacing is consistent across the (super-)family.
+    We may want all fonts within a super-family (all sibling families) to have the same vertical metrics so their line spacing is consistent across the super-family.
+
+    This is an experimental extended version of com.google.fonts/check/superfamily/vertical_metrics and for now it will only result in WARNs.
   """,
   misc_metadata = {
     'request': 'https://github.com/googlefonts/fontbakery/issues/1487'
@@ -733,7 +784,11 @@ def com_google_fonts_check_ttx_roundtrip(font):
 )
 def com_google_fonts_check_superfamily_vertical_metrics(superfamily_ttFonts):
   """Each font in a family must have the same vertical metrics values."""
-  failed = []
+  if len(superfamily_ttFonts) < 2:
+    yield SKIP, "Sibling families were not detected."
+    return
+
+  warn = []
   vmetrics = {
     "sTypoAscender": {},
     "sTypoDescender": {},
@@ -758,21 +813,16 @@ def com_google_fonts_check_superfamily_vertical_metrics(superfamily_ttFonts):
   for k, v in vmetrics.items():
     metric_vals = set(vmetrics[k].values())
     if len(metric_vals) != 1:
-      failed.append(k)
+      warn.append(k)
 
-  if len(family_ttFonts) > 1:
-    family_str = "super-family"
-  else:
-    family_str = "family"
-
-  if failed:
+  if warn:
     for k in failed:
       s = ["{}: {}".format(k, v) for k, v in vmetrics[k].items()]
       s = "\n".join(s)
-      yield FAIL, (f"{k} is not the same across the {family_str}:\n"
+      yield WARN, (f"{k} is not the same across the super-family:\n"
                    f"{s}")
   else:
-    yield PASS, f"Vertical metrics are the same across the {family_str}"
+    yield PASS, f"Vertical metrics are the same across the super-family."
 
 
 profile.auto_register(globals())